Add test for reverting change, and refactor
